+++ src/getFile.php
@@ -28,6 +28,8 @@ if (!$line) {
   die("ERR5 - file not found");
 }
 
+$accessGroupIds = [];
+
 //check user rights to download here:
 //if the user is logged in, he need to have the rights to
 //if agent provides his voucher, check it.
@@ -41,9 +43,10 @@ if (!$LOGIN->isLoggedin()) {
     else if ($apiKey->getStartValid() > time() || $apiKey->getEndValid() < time()) {
       die("Expired access key!");
     }
-    else if (!$this->hasPermission(USection::FILE, USectionFile::GET_FILE, $apiKey)) {
+    else if (!$ACCESS_CONTROL->hasPermission(USection::FILE, USectionFile::GET_FILE, $apiKey)) {
       die("Permission denied!");
     }
+    $accessGroupIds = Util::arrayOfIds(AccessUtils::getAccessGroupsOfUser($FACTORIES::getUserFactory()->get($apiKey->getUserId())));
   }
   else {
     $token = @$_GET['token'];
@@ -55,10 +58,18 @@ if (!$LOGIN->isLoggedin()) {
     if ($agent->getIsTrusted() < $line->getIsSecret()) {
       die("No access!");
     }
+    $accessGroupIds = Util::arrayOfIds(AccessUtils::getAccessGroupsOfAgent($agent));
   }
 }
 else if (!$ACCESS_CONTROL->hasPermission(DAccessControl::VIEW_FILE_ACCESS)) {
   die("No access!");
+}
+else{
+  $accessGroupIds = Util::arrayOfIds(AccessUtils::getAccessGroupsOfUser($LOGIN->getUser()));
+}
+
+if(!in_array($line->getAccessGroupId(), $accessGroupIds)){
+  die("Access denied to file because of access groups!");
 }
 
 $filename = dirname(__FILE__) . "/files/" . $line->getFilename();
@@ -91,12 +102,12 @@ header("Content-Description: " . $line->getFilename());
 header("Content-Disposition: attachment; filename=\"" . $line->getFilename() . "\"");
 
 if (isset($_SERVER['HTTP_RANGE'])) {
-  
+
   $c_start = $start;
   $c_end = $end;
-  
+
   list(, $range) = explode('=', $_SERVER['HTTP_RANGE'], 2);
-  
+
   if (strpos($range, ',') !== false) {
     header('HTTP/1.1 416 Requested Range Not Satisfiable');
     header("Content-Range: bytes $start-$end/$size");
@@ -135,7 +146,7 @@ header("Content-Length: " . $length);
 
 $buffer = 1024 * 100;
 while (!feof($fp) && ($p = ftell($fp)) <= $end) {
-  
+
   if ($p + $buffer > $end) {
     $buffer = $end - $p + 1;
   }